Extract Codex Connector review request decision
